Export MessageType enum as non-type
