+++ .changeset/shy-melons-look.md
@@ -1,0 +1,5 @@
+---
+"@great-detail/whatsapp": patch
+---
+
+Export MessageType enum as non-type
