+++ src/index.ts
@@ -18,6 +18,7 @@ export type * from "./types/Message/MessageMedia.js";
 export type * from "./types/Message/MessageReferral.js";
 export type * from "./types/Message/MessageSystem.js";
 export type * from "./types/Message/MessageText.js";
+export * from "./types/Message/MessageType.js";
 
 export type * from "./PhoneNumbers/index.js";
