+++ src/types/Message/index.ts
@@ -13,6 +13,7 @@ import { CreateMessageContact } from "./MessageContact.js";
 import { CreateMessageLocation } from "./MessageLocation.js";
 import { CreateMessageMedia } from "./MessageMedia.js";
 import { CreateMessageText } from "./MessageText.js";
+import { MessageType } from "./MessageType.js";
 
 /**
  * WhatsApp Message ID.
@@ -26,37 +27,6 @@ export type MessageRecipientType =
   | "individual"
   | "group"
   | (string & NonNullable<unknown>);
-
-export enum MessageType {
-  Audio = "audio", // for audio messages.
-  Contacts = "contacts", // for contact messages.
-  Document = "document", // for document messages.
-  Image = "image", // for image messages.
-  Interactive = "interactive", // for list and reply button messages.
-  Location = "location", // for location messages.
-  Reaction = "sticker", // for reaction messages.
-  // eslint-disable-next-line @typescript-eslint/no-duplicate-enum-values
-  Sticker = "sticker", // for sticker messages.
-
-  /**
-   * Not supported for Outgoing Messages.
-   */
-  System = "system",
-
-  /**
-   * Not supported for Outgoing Messages.
-   */
-  Button = "button",
-
-  /**
-   * Not supported for Outgoing Messages.
-   */
-  Order = "order",
-
-  Template = "template", // for template messages. Text and media (images and documents) message templates are supported.
-  Text = "text", // for text messages.
-  Video = "video", // for video messages
-}
 
 export type CreateMessageOptions = {
   phoneNumberID: PhoneNumberID;
